+++ test/controllers/codes-postaux/communes.js
@@ -1,12 +1,12 @@
 /* eslint-env node, mocha */
 const request = require('supertest');
-const server = require('../../../server');
+const app = require('../../../app');
 
 describe('/codes-postaux', function() {
     describe('/communes', function() {
         describe('too short postal code', function() {
             it('should reply with 400', function(done) {
-                request(server)
+                request(app)
                     .get('/api/codes-postaux/communes/7800')
                     .expect(400, done);
             });
@@ -14,7 +14,7 @@ describe('/codes-postaux', function() {
 
         describe('too long postal code', function() {
             it('should reply with 400', function(done) {
-                request(server)
+                request(app)
                     .get('/api/codes-postaux/communes/780000')
                     .expect(400, done);
             });
@@ -22,7 +22,7 @@ describe('/codes-postaux', function() {
 
         describe('alphabetic input', function() {
             it('should reply with 400', function(done) {
-                request(server)
+                request(app)
                     .get('/api/codes-postaux/communes/2A230')
                     .expect(400, done);
             });
@@ -30,7 +30,7 @@ describe('/codes-postaux', function() {
 
         describe('non-existent postal code', function() {
             it('should reply with 404', function(done) {
-                request(server)
+                request(app)
                     .get('/api/codes-postaux/communes/78123')
                     .expect(404, done);
             });
@@ -51,20 +51,20 @@ describe('/codes-postaux', function() {
             } ];
 
             it('should reply with 200', function(done) {
-                request(server)
+                request(app)
                     .get(`/api/codes-postaux/communes/${POSTAL_CODE}`)
                     .expect(200, done);
             });
 
             it('should reply with a JSON payload', function(done) {
-                request(server)
+                request(app)
                     .get(`/api/codes-postaux/communes/${POSTAL_CODE}`)
                     .expect('Content-Type', /json/)
                     .end(done);
             });
 
             it('should reply with an array', function(done) {
-                request(server)
+                request(app)
                     .get(`/api/codes-postaux/communes/${POSTAL_CODE}`)
                     .expect(PAYLOAD, done);
             });
