+++ test/controllers/gpu/all.js
@@ -2,13 +2,13 @@
 const request = require('supertest');
 const expect = require('expect.js');
 
-const server = require('../../../server');
+const app = require('../../../app');
 
 describe('/api/gpu/all', function() {
 
     describe('with invalid geom', function() {
         it('should reply an error', function(done) {
-            request(server)
+            request(app)
                 .get('/api/gpu/all?geom={"type":"Point","coordinates":[1.654399]}')
                 .expect(400)
                 .end(done);
@@ -20,7 +20,7 @@ describe('/api/gpu/all', function() {
     describe('with point at [1.654399,48.112235] (Rennes)', function() {
         it('should reply a list of FeatureCollection', function(done) {
             // en attente de résolution problème de type sur assiette-sup-p
-            request(server)
+            request(app)
                 .get('/api/gpu/all?geom={"type":"Point","coordinates":[1.654399,48.112235]}')
                 .expect(200)
                 .expect(res => {
