+++ test/controllers/gpu/document.js
@@ -2,14 +2,14 @@
 const request = require('supertest');
 const expect = require('expect.js');
 
-const server = require('../../../server');
+const app = require('../../../app');
 
 describe('/api/gpu/document', function() {
     
 
     describe('with point at [1.654399,48.112235] (Rennes)', function() {
         it('should reply a FeatureCollection containing a valid Feature', function(done) {
-            request(server)
+            request(app)
                 .get('/api/gpu/document?geom={"type":"Point","coordinates":[1.654399,48.112235]}')
                 .expect(200)
                 .expect(res => {
