+++ test/controllers/gpu/zone-urba.js
@@ -2,13 +2,13 @@
 const request = require('supertest');
 const expect = require('expect.js');
 
-const server = require('../../../server');
+const app = require('../../../app');
 
 describe('/api/gpu/zone-urba', function() {
 
     describe('with point at [1.654399,48.112235] (Rennes)', function() {
         it('should reply a FeatureCollection containing a valid Feature', function(done) {
-            request(server)
+            request(app)
                 .get('/api/gpu/zone-urba?geom={"type":"Point","coordinates":[1.654399,48.112235]}')
                 .expect(200)
                 .expect(res => {
