+++ test/controllers/rpg/rpgv1.js
@@ -2,12 +2,12 @@
 const request = require('supertest');
 const expect = require('expect.js');
 
-const server = require('../../../server');
+const app = require('../../../app');
 
 describe('/api/rpg/v1', function() {
     describe('without filtering parameter', function() {
         it('should reply with 400', function(done) {
-            request(server)
+            request(app)
                 .get('/api/rpg/v1')
                 .expect(400, done);
         });
@@ -15,7 +15,7 @@ describe('/api/rpg/v1', function() {
 
     describe('with invalid geom', function() {
         it('should reply an error', function(done) {
-            request(server)
+            request(app)
                 .get('/api/rpg/v1?geom={"type":"Point","coordinates":[1.654399]}')
                 .expect(400)
                 .end(done);
@@ -27,7 +27,7 @@ describe('/api/rpg/v1', function() {
     /*describe('with point at [1.654399,48.112235] (Rennes)', function() {
         it('should reply a list of FeatureCollection', function(done) {
             // en attente de résolution problème de type sur assiette-sup-p
-            request(server)
+            request(app)
                 .get('/api/rpg/v1?annee=2013&geom={"type":"Point","coordinates":[1.654399,48.112235]}')
                 .expect(200)
                 .expect(res => {
